moved contract tests into submods
